+++ .gitignore
@@ -147,3 +147,4 @@ output/
 
 # Mac files
 .DS_Store
+*.dot
