+++ stirling-engine.code-workspace
@@ -7,37 +7,39 @@
 	],
 	"settings": {
 		"workbench.colorCustomizations": {
-			"activityBar.activeBackground": "#4ff78a",
+			"activityBar.activeBackground": "#1f644b",
 			"activityBar.activeBorder": "#5858c7",
-			"activityBar.background": "#4ff78a",
-			"activityBar.foreground": "#15202b",
-			"activityBar.inactiveForeground": "#15202b99",
-			"activityBarBadge.background": "#9a67f8",
-			"activityBarBadge.foreground": "#15202b",
-			"editorGroup.border": "#4ff78a",
-			"panel.border": "#4ff78a",
-			"sash.hoverBorder": "#4ff78a",
-			"sideBar.border": "#4ff78a",
-			"statusBar.background": "#1ef56a",
-			"statusBar.border": "#1ef56a",
-			"statusBar.foreground": "#15202b",
-			"statusBarItem.hoverBackground": "#0ad652",
-			"statusBarItem.remoteBackground": "#1ef56a",
-			"statusBarItem.remoteForeground": "#15202b",
-			"tab.activeBorder": "#4ff78a",
-			"titleBar.activeBackground": "#1ef56a",
-			"titleBar.activeForeground": "#15202b",
-			"titleBar.border": "#1ef56a",
-			"titleBar.inactiveBackground": "#1ef56a99",
-			"titleBar.inactiveForeground": "#15202b99",
-			"commandCenter.border": "#15202b99",
-			"statusBar.debuggingBackground": "#f51ea9",
-			"statusBar.debuggingBorder": "#f51ea9",
+			"activityBar.background": "#1f644b",
+			"activityBar.foreground": "#e7e7e7",
+			"activityBar.inactiveForeground": "#e7e7e799",
+			"activityBarBadge.background": "#371749",
+			"activityBarBadge.foreground": "#e7e7e7",
+			"editorGroup.border": "#1f644b",
+			"panel.border": "#1f644b",
+			"sash.hoverBorder": "#1f644b",
+			"sideBar.border": "#1f644b",
+			"statusBar.background": "#133d2e",
+			"statusBar.border": "#133d2e",
+			"statusBar.foreground": "#e7e7e7",
+			"statusBarItem.hoverBackground": "#1f644b",
+			"statusBarItem.remoteBackground": "#133d2e",
+			"statusBarItem.remoteForeground": "#e7e7e7",
+			"tab.activeBorder": "#1f644b",
+			"titleBar.activeBackground": "#133d2e",
+			"titleBar.activeForeground": "#e7e7e7",
+			"titleBar.border": "#133d2e",
+			"titleBar.inactiveBackground": "#133d2e99",
+			"titleBar.inactiveForeground": "#e7e7e799",
+			"commandCenter.border": "#e7e7e799",
+			"statusBar.debuggingBackground": "#3d1322",
+			"statusBar.debuggingBorder": "#3d1322",
 			"statusBar.debuggingForeground": "#e7e7e7"
 		},
-		"peacock.color": "#1ef56a",
+		"peacock.color": "#133d2e",
 		"cSpell.words": [
-			"audiowaveform"
+			"audiowaveform",
+			"ffprobe",
+			"Unparser"
 		],
 		"blocksort.enableCodeActions": "*",
 		"VSCodeCounter.exclude": ["docs/*"]
